update jwt config

+++ config/jwt.php
@@ -7,8 +7,8 @@ return [
         'private' => env('JWT_PRIVATE_KEY'),
         'passphrase' => env('JWT_PASSPHRASE'),
     ],
-    'ttl' => env('JWT_TTL', 60),
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    'ttl' => (int) env('JWT_TTL', 60), // تبدیل به integer
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160), // تبدیل به integer
     'algo' => env('JWT_ALGO', 'HS256'),
     'required_claims' => [
         'iss',
@@ -23,9 +23,9 @@ return [
         // 'bar',
     ],
     'lock_subject' => true,
-    'leeway' => env('JWT_LEEWAY', 0),
+    'leeway' => (int) env('JWT_LEEWAY', 0), // تبدیل به integer
     'blacklist_enabled' => env('JWT_BLACKLIST_ENABLED', true),
-    'blacklist_grace_period' => env('JWT_BLACKLIST_GRACE_PERIOD', 0),
+    'blacklist_grace_period' => (int) env('JWT_BLACKLIST_GRACE_PERIOD', 0), // تبدیل به integer
     'decrypt_cookies' => false,
     'providers' => [
         'jwt' => Tymon\JWTAuth\Providers\JWT\Lcobucci::class,
